Src/GCS_BIGQUERY_GGFUNCTION/requirements.txt
# Function dependencies, for example:
# package>=version
google.cloud.storage
google.cloud.bigquery